pyproject.toml
[project]
name = "azure-governance-platform"
version = "2.5.0"
description = "Multi-tenant Azure governance platform for cost optimization, compliance monitoring, resource management, and identity governance"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Cloud Governance Team"}
]
license = {text = "MIT"}
keywords = ["azure", "governance", "multi-tenant", "cost-management", "compliance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.44.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.49",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.26",
    "httpx>=0.26.0",
    "apscheduler>=3.10.0",
    # Authentication & Security
    "PyJWT[crypto]>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=46.0.7",
    # Azure SDKs
    "requests>=2.33.1",  # CVE-2026-25645 fix (transitive via azure-identity/msal)
    "azure-identity>=1.25.3",
    "azure-mgmt-resource>=23.0.0,<24.0.0",
    "azure-mgmt-costmanagement>=4.0.0",
    "azure-mgmt-policyinsights>=1.0.0",
    "azure-mgmt-security>=5.0.0",
    "azure-mgmt-authorization>=4.0.0",
    "azure-mgmt-subscription>=3.1.1",
    "azure-keyvault-secrets>=4.7.0",  # Key Vault secret storage
    "six>=1.17.0",
    "alembic>=1.18.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",  # JSON Schema validation for custom compliance rules (CM-002)
    # Token blacklist & caching (optional but required for production)
    "redis>=7.4.0",
    # Azure SQL / ODBC support
    "pyodbc>=4.0.39",
    "opentelemetry-api>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "opentelemetry-exporter-otlp>=1.41.0",
    "azure-mgmt-compute>=37.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=7.1.0",
    "httpx>=0.26.0",
    "ruff>=0.15.10",
    "mypy>=1.20.1",
    "pre-commit>=3.6.0",
]

[project.urls]
"Homepage" = "https://github.com/htt-brands/azure-governance-platform"
"Bug Tracker" = "https://github.com/htt-brands/azure-governance-platform/issues"

[project.scripts]
govern = "app.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "RUF100", # flag unused/stale noqa directives (prevents noqa-rot)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "E741",  # ambiguous variable name (l in HSL is standard)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"scripts/*" = ["E402"]
"tests/unit/test_*.py" = ["E402"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --ignore=tests/e2e --ignore=tests/load --ignore=tests/smoke --ignore=tests/staging"
markers = [
    "azure_creds_required: marks tests requiring Azure credentials",
    "auth_required: marks tests requiring authentication",
    "slow: marks tests as slow",
    "e2e: marks tests as end-to-end (require Playwright and running server)",
    "smoke: marks tests as smoke tests (require running server)",
    "staging: marks tests as staging validation (require live staging URL)",
    "visual: marks visual-parity tests (require Playwright browsers + pinned baselines)",
]

[dependency-groups]
dev = [
    "aiosmtplib>=5.1.0",
    "locust>=2.43.4",
    "mypy>=1.19.1",
    "pillow>=10.0.0",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.3",
]

[[tool.mypy.overrides]]
module = [
    "app.core.retry",
    "app.core.circuit_breaker",
    "app.core.cache",
    "app.core.monitoring",
    "app.core.notifications",
    "app.core.rate_limit",
    "app.core.database",
    "app.core.auth",
    "app.core.resilience",
]
ignore_errors = true

[tool.mutmut]
paths_to_mutate = "app/"
backup = false
runner = "python -m pytest tests/ -x"
tests_dir = "tests/"
